move statisticalmodel io classes to separate subpackage
