+++ src/test/scala/scalismo/io/statisticalmodel/StatismoDomainIOTests.scala
@@ -13,28 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package scalismo.io
+package scalismo.io.statisticalmodel
 
 import scalismo.ScalismoTestSuite
-import scalismo.geometry._
-import scalismo.common.{PointId}
+import scalismo.common.PointId
 import scalismo.common.UnstructuredPoints.Create.{CreateUnstructuredPoints2D, CreateUnstructuredPoints3D}
-import scalismo.mesh.{
-  LineCell,
-  LineList,
-  LineMesh2D,
-  LineMesh3D,
-  TetrahedralCell,
-  TetrahedralList,
-  TetrahedralMesh3D,
-  TriangleCell,
-  TriangleList,
-  TriangleMesh2D,
-  TriangleMesh3D
-}
+import scalismo.geometry._
+import scalismo.io.StatismoDomainIO
+import scalismo.mesh._
 
 import scala.language.higherKinds
-import scala.util.{Failure, Success, Try}
 
 class StatismoDomainIOTests extends ScalismoTestSuite {
